+++ bin/check-kube-pods-pending.rb
@@ -28,6 +28,8 @@
 #     --token-file TOKEN-FILE      File containing bearer token for authorization
 # -n NAMESPACES,                   Exclude the specified list of namespaces
 #     --exclude-namespace
+# -i NAMESPACES,                   Include the specified list of namespaces, an
+#     --include-namespace          empty list includes all namespaces
 # -t, --timeout TIMEOUT            Threshold for pods to be in the pending state
 # -f, --filter FILTER              Selector filter for pods to be checked
 # -p, --pods PODS                  Optional list of pods to check.
@@ -73,6 +75,13 @@ class AllPodsAreReady < Sensu::Plugins::Kubernetes::CLI
          proc: proc { |a| a.split(',') },
          default: ''
 
+  option :include_namespace,
+         description: 'Include the specified list of namespaces',
+         short: '-i NAMESPACES',
+         long: '--include-namespace',
+         proc: proc { |a| a.split(',') },
+         default: ''
+
   def run
     pods_list = []
     failed_pods = []
@@ -89,7 +98,7 @@ class AllPodsAreReady < Sensu::Plugins::Kubernetes::CLI
     end
     pods.each do |pod|
       next if pod.nil?
-      next if config[:exclude_namespace].include?(pod.metadata.namespace)
+      next if should_exclude_namespace(pod.metadata.namespace)
       next unless pods_list.include?(pod.metadata.name) || pods_list.include?('all')
       # Check for pending state
       next unless pod.status.phase == 'Pending'
@@ -112,5 +121,10 @@ class AllPodsAreReady < Sensu::Plugins::Kubernetes::CLI
     return list.split(',') if list && list.include?(',')
     return [list] if list
     ['']
+  end
+
+  def should_exclude_namespace(namespace)
+    return !config[:include_namespace].include?(namespace) unless config[:include_namespace].empty?
+    config[:exclude_namespace].include?(namespace)
   end
 end